+++ src/lib/Components/Inbox/Conversations/__tests__/Avatar-tests.tsx
@@ -1,0 +1,11 @@
+import "jest-snapshots-svg"
+import * as React from "react"
+import "react-native"
+import * as renderer from "react-test-renderer"
+
+import Avatar from "../Avatar"
+
+it("looks correct for a user with two initials", () => {
+  const avatar = renderer.create(<Avatar isUser={true} initials={"MC"} />).toJSON()
+  expect(avatar).toMatchSnapshot()
+})