+++ src/lib/Containers/Conversation.tsx
@@ -63,15 +63,18 @@ const ComposerContainer = styled.View`
 
 export class Conversation extends React.Component<RelayProps, any> {
   renderMessage({ item }) {
-    const artwork = this.props.me.conversation.artworks[0]
-    const conversation = this.props.me.conversation
+    const me = this.props.me
+    const artwork = me.conversation.artworks[0]
+    const conversation = me.conversation
     const partnerName = conversation.to.name
     const senderName = item.is_from_user ? conversation.from.name : partnerName
+    const initials = item.is_from_user ? conversation.from.initials : conversation.to.initials
 
     return (
       <Message
         message={item}
         senderName={senderName}
+        initials={initials}
         artworkPreview={
           item.first_message &&
           <ArtworkPreview
@@ -127,9 +130,11 @@ export default Relay.createContainer(Conversation, {
           from {
             name
             email
+            initials
           }
           to {
             name
+            initials
           }
           messages(first: 10) {
             pageInfo {
@@ -138,6 +143,7 @@ export default Relay.createContainer(Conversation, {
             edges {
               node {
                 id
+                is_from_user
                 ${Message.getFragment("message")}
               }
             }
@@ -154,14 +160,17 @@ export default Relay.createContainer(Conversation, {
 
 interface RelayProps {
   me: {
+    initials: string
     conversation: {
       id: string
       from: {
         name: string
         email: string
+        initials: string
       }
       to: {
         name: string
+        initials: string
       }
       artworks: any[]
       messages: {